Move BazelFileSystemModule into bazel package
Currently, it is included in runtime library with glob which is shared between bazel and blaze. Moving it to bazel package to prevent from packaging it when building blaze.

An upcoming change (in https://github.com/bazelbuild/bazel/pull/18784) will package BLAKE3 to this module and we don't want it in blaze.

PiperOrigin-RevId: 547516042
Change-Id: I673fa3d6824d56c85e85b02d13b4eb56571edda9

(cherry picked from commit 6d6bbdcd7ac7e8130741417d891e641495b046b2)